### 2020-06-08 #2312 @arfon
:wave: @sambit-giri - thanks for your submission to JOSS. Can you tell me if this submission accompanies a paper in a AAS journal?
