### 2020-06-08 #2312 @arfon
Also, from what I can find, there's no `paper.md` (or other JOSS-related materials) in this repository. Please familiarize yourself with our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html) and let us know when you've added your paper.
