+++ paper/paper.bib
@@ -1,0 +1,212 @@
+@phdthesis{giri2019tomographic,
+  title={Tomographic studies of the 21-cm signal during reionization: Going beyond the power spectrum},
+  author={Giri, Sambit K},
+  year={2019},
+  school={Department of Astronomy, Stockholm University}
+}
+
+@article{Furlanetto2006CosmologyUniverse,
+    title = {{Cosmology at low frequencies: The 21 cm transition and the high-redshift Universe}},
+    year = {2006},
+    author = {Furlanetto, S.~R. and Oh, S.~P. and Briggs, F.~H.},
+    month = {10},
+    pages = {181--301},
+    volume = {433},
+    doi = {10.1016/j.physrep.2006.08.002}
+}
+
+@misc{Pritchard201221cmCentury,
+    title = {{21cm cosmology in the 21st century}},
+    year = {2012},
+    booktitle = {Reports on Progress in Physics},
+    author = {Pritchard, Jonathan R. and Loeb, Abraham},
+    number = {8},
+    month = {8},
+    pages = {86901},
+    volume = {75},
+    doi = {10.1088/0034-4885/75/8/086901},
+    issn = {00344885},
+    pmid = {22828208},
+    arxivId = {1109.6012}
+}
+
+@article{Datta2012Light-coneSpectrum,
+    title = {{Light-cone effect on the reionization 21-cm power spectrum}},
+    year = {2012},
+    journal = {Monthly Notices of the Royal Astronomical Society},
+    author = {Datta, Kanan K. and Mellema, Garrelt and Mao, Yi and Iliev, Ilian T. and Shapiro, Paul R. and Ahn, Kyungjin},
+    number = {3},
+    month = {8},
+    pages = {1877--1891},
+    volume = {424},
+    publisher = {Oxford University Press},
+    url = {},
+    doi = {10.1111/j.1365-2966.2012.21293.x},
+    issn = {00358711}
+}
+
+@article{Giri2018BubbleTomography,
+    title = {{Bubble size statistics during reionization from 21-cm tomography}},
+    year = {2018},
+    journal = {MNRAS},
+    author = {Giri, Sambit K and Mellema, Garrelt and Dixon, Keri L and Iliev, Ilian T},
+    pages = {2949--2964},
+    volume = {473},
+    url = {},
+    doi = {10.1093/mnras/stx2539},
+    arxivId = {1706.00665},
+    keywords = {first stars – early Universe – radio lines, galaxies, image processing – dark ages, methods, reionization, statistical – techniques}
+}
+
+@article{Jensen2013ProbingDistortions,
+    title = {{Probing reionization with LOFAR using 21-cm redshift space distortions}},
+    year = {2013},
+    journal = {Monthly Notices of the Royal Astronomical Society},
+    author = {Jensen, Hannes and Datta, Kanan K. and Mellema, Garrelt and Chapman, Emma and Abdalla, Filipe B. and Iliev, Ilian T. and Mao, Yi and Santos, Mario G. and Shapiro, Paul R. and Zaroubi, Saleem and Bernardi, G. and Brentjens, M. A. and de Bruyn, A. G. and Ciardi, B. and Harker, G. J. A. and Jeli{\'{c}}, V. and Kazemi, S. and Koopmans, L. V. E. and Labropoulos, P. and Martinez, O. and Offringa, A. R. and Pandey, V. N. and Schaye, J. and Thomas, R. M. and Veligatla, V. and Vedantham, H. and Yatawatta, S.},
+    number = {1},
+    month = {10},
+    pages = {460--474},
[…]
