+++ paper/README.md
@@ -1,0 +1,8 @@
+# Paper submission
+
+Tools21cm paper for submission to the [Journal of Open Source Science](https://joss.theoj.org/).
+
+[Paper submission instructions](https://joss.readthedocs.io/en/latest/submitting.html) -
+this link also includes formatting instructions.
+
+[Review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html)
