### 2020-06-09 #2312 @sambit-giri
@arfon There is no accompanying a paper in an AAS journal.
We have put the paper for submission in the [repository](https://github.com/sambit-giri/tools21cm/tree/master/paper).
I apologise for the inconvenience.
