+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @phdthesis{giri2019tomographic,
   title={Tomographic studies of the 21-cm signal during reionization: Going beyond the power spectrum},
-  author={Giri, Sambit K},
+  author={Giri, Sambit K.},
   year={2019},
   school={Department of Astronomy, Stockholm University}
 }
@@ -49,7 +49,7 @@
     title = {{Bubble size statistics during reionization from 21-cm tomography}},
     year = {2018},
     journal = {MNRAS},
-    author = {Giri, Sambit K and Mellema, Garrelt and Dixon, Keri L and Iliev, Ilian T},
+    author = {Giri, Sambit K. and Mellema, Garrelt and Dixon, Keri L and Iliev, Ilian T},
     pages = {2949--2964},
     volume = {473},
     url = {},
@@ -103,7 +103,7 @@
     title = {{Neutral island statistics during reionization from 21-cm tomography}},
     year = {2019},
     journal = {Monthly Notices of the Royal Astronomical Society},
-    author = {Giri, Sambit K and Mellema, Garrelt and Aldheimer, Thomas and Dixon, Keri L and Iliev, Ilian T},
+    author = {Giri, Sambit K. and Mellema, Garrelt and Aldheimer, Thomas and Dixon, Keri L and Iliev, Ilian T},
     number = {2},
     month = {10},
     pages = {1590--1605},
@@ -133,7 +133,7 @@
     title = {{Optimal identification of H ii regions during reionization in 21-cm observations}},
     year = {2018},
     journal = {Monthly Notices of the Royal Astronomical Society},
-    author = {Giri, Sambit K and Mellema, Garrelt and Ghara, Raghunath},
+    author = {Giri, Sambit K. and Mellema, Garrelt and Ghara, Raghunath},
     number = {4},
     month = {10},
     pages = {5596--5611},
