+++ paper/paper.bib
@@ -99,6 +99,18 @@
     arxivId = {1607.06282}
 }
 
+@article{Ross2019EvaluatingDawn,
+    title = {{Evaluating the QSO contribution to the 21-cm signal from the Cosmic Dawn}},
+    year = {2019},
+    journal = {MNRAS},
+    author = {Ross, Hannah E. and Dixon, Keri L and Ghara, Raghunath and Iliev, Ilian T and Mellema, Garrelt},
+    pages = {1101--1119},
+    volume = {487},
+    url = {http://eor.berkeley.edu/.},
+    doi = {10.1093/mnras/stz1220},
+    keywords = {first stars-galaxies: quasars: general-intergalactic medium-cosmology: theory-cosmology: large-scale struc-ture of Universe, radiative transfer-cosmology: dark ages, reionizaion}
+}
+
 @article{Giri2019NeutralTomography,
     title = {{Neutral island statistics during reionization from 21-cm tomography}},
     year = {2019},
