### 2020-06-09 #2312 @pibion
@arfon is it okay that my expertise is squarely in the "particle" part of particle astrophysics?  I do have some basic familiarity with this topic and am happy to review, but we should definitely have someone on the team who is expert in this kind of analysis.

@sambit-giri do you have any reviewers you could recommend?
