### 2020-06-09 #2312 @arfon
> @arfon is it okay that my expertise is squarely in the "particle" part of particle astrophysics? I do have some basic familiarity with this topic and am happy to review, but we should definitely have someone on the team who is expert in this kind of analysis.

Yeah, I think this is OK :-) . Let's just make sure we get some expert reviewers as you say.

@pibion - The next step here is to ask Whedon to assign this paper to you. You can do either `@whedon assign me as editor` or `@whedon assign @pibion as editor`.
