### 2020-06-10 6290079
removed the /2 factor
