+++ t2c/telescope_functions.py
@@ -161,7 +161,7 @@ def get_uv_coverage(Nbase, z, ncells, boxsize=None):
 	if not boxsize: boxsize = conv.LB
 	uv_map = np.zeros((ncells,ncells))
 	theta_max = boxsize/cm.z_to_cdist(z)
-	Nb  = np.round(Nbase*theta_max/2)
+	Nb  = np.round(Nbase*theta_max)
 	Nb  = Nb[(Nb[:,0]<ncells/2)]
 	Nb  = Nb[(Nb[:,1]<ncells/2)]
 	Nb  = Nb[(Nb[:,2]<ncells/2)]
