+++ t2c/telescope_functions.py
@@ -173,7 +173,7 @@ def get_uv_coverage(Nbase, z, ncells, boxsize=None):
 	for p in range(xx.shape[0]): 
 		uv_map[int(xx[p]),int(yy[p])] += 1
 		uv_map[-int(xx[p]),-int(yy[p])] += 1
-	return uv_map
+	return uv_map/2
 
 
 def kanan_noise_image_ska(z, uv_map, depth_mhz, obs_time, int_time, N_ant_ska=564., verbose=True):
