+++ docs/readme.rst
@@ -2,7 +2,7 @@
 tools21cm
 =========
 
-A python package for analysing 21-cm signals from the Epoch of Reionization (EoR) and Cosmic Dawn (CD).
+A python package for analysing 21-cm signals from the Epoch of Reionization (EoR) and Cosmic Dawn (CD). The source files can be found [here](https://github.com/sambit-giri/tools21cm).
 
 Note: There are some modules in the package that are still under active development. Therefore please contact the authors if you get erronous results.
 
