### 2020-06-18 #2363 @whedon
[REVIEW]: Tools21cm: A python package to analyse the large-scale 21-cm signal from the Epoch of Reionization and Cosmic Dawn
**Submitting author:** @sambit-giri (<a href="http://orcid.org/0000-0002-2560-536X">Sambit Kumar Giri</a>)
**Repository:** <a href="https://github.com/sambit-giri/tools21cm" target ="_blank">https://github.com/sambit-giri/tools21cm</a>
**Version:** v2.1

@nicholebarry & @aureliocarnero & @ziotom78 , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
