### 2020-07-03 #2363 @pibion
@sambit-giri, @aureliocarnero on the authorship issue: JOSS does need an affiliation for every author, but this is only required to make the latex template happy.  So for someone without an academic affiliation, they can put whatever seems most appropriate.  For example,

```
authors:
  - name: Jane Doe
    affiliation: 1
affiliations:
 - name: None
   index: 1
```

Would show "None" as the affiliation.

The author can put whatever seems most appropriate here, though.  In addition to the "none" option, another suggestion is"Self."
