### 2020-07-03 #2363 @aureliocarnero
Dear @sambit-giri please see message above. I would recommend to include Hannes as author to the paper if he agrees.
I have some comments on the references in the paper draft.
Since you reference several papers of yourself from 2019, could you please put 2019a, 2019b... etcerera? The same as you do with your 2018 papers.
Also, there is one reference: Ross et al., 2017, that at some point you reference as: Ross, Dixon, Iliev,
& Mellema, 2017.

Also, I'm not sure when you should use et al, or the list of names as a function of the numbers of authors. I think is 3, if there are only 3 authors, you can place the three names, but if they are more than 3, then you should put et al. Homogenize it through the text, please. @pibion , could you please confirm this issue? When we should use et al, and when the full list of authors?
