### 2020-07-03 #2363 @aureliocarnero
@sambit-giri also, one of the checks I need to do related to the paper is:
State of the field: Do the authors describe how this software compares to other commonly-used packages?

You actually say nothing about the state of the field, are there any other software that compares? In these terms, I think is important you name in the draft what you commented above: "We modified  c2raytools (https://github.com/hjens/c2raytools) package and merged it into tools21cm. ", referencing the code appropietly.
Cheers
Aurelio
