### 2020-07-06 #2363 @pibion
@nicholebarry, @ziotom78 this is a reminder that we're asking our reviewers to aim to complete their checklist in about six weeks.  If you have an idea of what your time frame is for getting started, you can let me know and I can set a reminder with `whedon`.
