### 2020-07-07 #2363 @aureliocarnero
@sambit-giri another question about installation. It seems like it is installed properly with the instructions you give, but I have a question. When I install it with python setup install, it appears a warning message as follows:
UserWarning: Unknown distribution option: 'install_requires'.
Im using python3, do you know if this can cause a problem to other users? Please verify the meaning of this warning.
Installing it with pip works perfectly, though
