### 2020-07-07 #2363 @garrelt
Dear Aurelio,

Although individual researchers have undoubtedly developed their own tools to perform some of the calculations
included in tools21cm, we are not aware of any open source package similar to tools21cm. We did search both
with google and on github, but did not find anything.

c2raytools was a set of python routines developed by Hannes Jensen and me to analyse reionization
simulations performed with the C2-Ray radiative transfer code. This set of routines is no longer
maintained and never had a user base beyond our nearest collaborators. It formed the basis
of the tools21cm package, which is however more general, hence the name change.

I hope this answers your query.

Best wishes,
Garrelt
> @sambit-giri also, one of the checks I need to do related to the paper is:
> State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
> You actually say nothing about the state of the field, are there any other software that compares? In these terms, I think is important you name in the draft what you commented above: "We modified c2raytools (https://github.com/hjens/c2raytools) package and merged it into tools21cm. ", referencing the code appropietly.
> Cheers
> Aurelio
