### 2020-07-08 #2363 @aureliocarnero
Dear @garrelt thank you for the clarification. No need to put anything in the text, then. Cheers
Aurelio
