### 2020-07-09 2a10d02
neighbors -> connectivity
