### 2020-07-09 e0e3099
SPA speed up
