+++ docs/tutorials.rst
@@ -8,4 +8,4 @@ The following examples will help you get started with ``tools21cm``:
 .. toctree::
    :maxdepth: 2
 
-   examples/read_write_plot
+   examples/tutorials
