### 2020-07-10 #2363 @sambit-giri
Dear @aureliocarnero
In order to test the code, some simulation output is required. We have now provided some data at https://ttt.astro.su.se/~gmell/244Mpc/TestData_tools21cm/
The link contains data along with a README explaining it.

You can test the code on this data by running the tutorial given at https://tools21cm.readthedocs.io/examples/tutorials.html

> Dear @sambit-giri , Im starting to test the code. One important aspect is that you need to provide automated tests with your code, but I cannot find those or how to run. According to the checklist, I need to verify that: "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"
> I cannot verify this at the moment
