### 2020-07-13 #2363 @nicholebarry
@sambit-giri,

The instrument papers for the MWA are [Tingay et al. 2013](https://ui.adsabs.harvard.edu/abs/2013PASA...30....7T/abstract) and [Wayth et al. 2018](https://ui.adsabs.harvard.edu/abs/2018PASA...35...33W/abstract) (Phase I and Phase II respectively). I suggest you cite those instead of Lonsdale et al. 2009.
