+++ paper/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 The Cosmic Dawn (CD) and Epoch of Reionization (EoR) are among the least understood epochs from the history of our Universe. Large-scale cosmological simulations are useful to understand this era and help prepare for future observations. `Tools21cm` is a data manipulation and analysis package for studying such cosmological simulations. It is written in the python programming language and designed to be very user-friendly.
 
 The 21 cm signal, produced by the spin-flip transition of neutral hydrogen, is a unique tracer of matter in large-scale structures present during the EoR and CD [e.g. @Furlanetto2006CosmologyUniverse; @Pritchard201221cmCentury]. This signal will be cosmologically redshifted and thus found at low radio frequencies.
-Much of the functionality of `Tools21cm` is focused on the construction and analysis of mock 21 cm observations in the context of current and upcoming radio telescopes, such as the Low Frequency Array [LOFAR; @vanHaarlem2013LOFAR:ARray], the Murchison Widefield Array [MWA; @Lonsdale2009TheOverview], Hydrogen Epoch of Reionization Array [HERA; @deboer2017hydrogen] and the Square Kilometre Array [SKA; @Mellema2013ReionizationArray], using cosmological simulations.
+Much of the functionality of `Tools21cm` is focused on the construction and analysis of mock 21 cm observations in the context of current and upcoming radio telescopes, such as the Low Frequency Array [LOFAR; @vanHaarlem2013LOFAR:ARray], the Murchison Widefield Array [MWA; @Tingay2013MWA, @Wayth2018MWAdesign], Hydrogen Epoch of Reionization Array [HERA; @deboer2017hydrogen] and the Square Kilometre Array [SKA; @Mellema2013ReionizationArray], using cosmological simulations.
 
 Radio telescopes typically observe the redshifted 21 cm signal over a range of frequencies and therefore produce 21 cm images at different redshifts. A sequence of such images from different redshifts is known as a tomographic data set and is three dimensional. `Tools21cm` can construct such tomographic data sets from simulation snapshots [@Datta2012Light-coneSpectrum; @Giri2018BubbleTomography]. When constructing these data sets, it can also add the impact of peculiar velocities, leading to an effect known as redshift space distortions [e.g. @Jensen2013ProbingDistortions; @Jensen2016TheMeasurements; @Giri2018BubbleTomography]. See @giri2019tomographic for a detailed description of tomographic 21 cm data sets.
 
