+++ paper/paper.bib
@@ -211,6 +211,68 @@
     keywords = {dark ages, reionization, first stars, instrumentation: interferometers, radio continuum: general, radio lines: general, telescopes}
 }
 
+@ARTICLE{Wayth2018MWAdesign,
+       author = {{Wayth}, Randall B. and {Tingay}, Steven J. and {Trott}, Cathryn M. and
+         {Emrich}, David and {Johnston-Hollitt}, Melanie and {McKinley}, Ben and
+         {Gaensler}, B.~M. and {Beardsley}, A.~P. and {Booler}, T. and
+         {Crosse}, B. and {Franzen}, T.~M.~O. and {Horsley}, L. and
+         {Kaplan}, D.~L. and {Kenney}, D. and {Morales}, M.~F. and {Pallot}, D. and
+         {Sleap}, G. and {Steele}, K. and {Walker}, M. and {Williams}, A. and
+         {Wu}, C. and {Cairns}, Iver. H. and {Filipovic}, M.~D. and
+         {Johnston}, S. and {Murphy}, T. and {Quinn}, P. and
+         {Staveley-Smith}, L. and {Webster}, R. and {Wyithe}, J.~S.~B.},
+        title = "{The Phase II Murchison Widefield Array: Design overview}",
+      journal = {\pasa},
+     keywords = {early universe, instrumentation: interferometers, radio continuum: general, radio lines: general, techniques: interferometric, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2018,
+        month = nov,
+       volume = {35},
+        pages = {33},
+          doi = {10.1017/pasa.2018.37},
+archivePrefix = {arXiv},
+       eprint = {1809.06466},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018PASA...35...33W},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Tingay2013MWA,
+       author = {{Tingay}, S.~J. and {Goeke}, R. and {Bowman}, J.~D. and {Emrich}, D. and
+         {Ord}, S.~M. and {Mitchell}, D.~A. and {Morales}, M.~F. and
+         {Booler}, T. and {Crosse}, B. and {Wayth}, R.~B. and {Lonsdale}, C.~J. and
+         {Tremblay}, S. and {Pallot}, D. and {Colegate}, T. and {Wicenec}, A. and
+         {Kudryavtseva}, N. and {Arcus}, W. and {Barnes}, D. and {Bernardi}, G. and
+         {Briggs}, F. and {Burns}, S. and {Bunton}, J.~D. and {Cappallo}, R.~J. and
+         {Corey}, B.~E. and {Deshpande}, A. and {Desouza}, L. and
+         {Gaensler}, B.~M. and {Greenhill}, L.~J. and {Hall}, P.~J. and
+         {Hazelton}, B.~J. and {Herne}, D. and {Hewitt}, J.~N. and
+         {Johnston-Hollitt}, M. and {Kaplan}, D.~L. and {Kasper}, J.~C. and
+         {Kincaid}, B.~B. and {Koenig}, R. and {Kratzenberg}, E. and
+         {Lynch}, M.~J. and {Mckinley}, B. and {Mcwhirter}, S.~R. and
+         {Morgan}, E. and {Oberoi}, D. and {Pathikulangara}, J. and {Prabu}, T. and
+         {Remillard}, R.~A. and {Rogers}, A.~E.~E. and {Roshi}, A. and
+         {Salah}, J.~E. and {Sault}, R.~J. and {Udaya-Shankar}, N. and
+         {Schlagenhaufer}, F. and {Srivani}, K.~S. and {Stevens}, J. and
[…]
