+++ paper/paper.bib
@@ -222,7 +222,7 @@
          {Johnston}, S. and {Murphy}, T. and {Quinn}, P. and
          {Staveley-Smith}, L. and {Webster}, R. and {Wyithe}, J.~S.~B.},
         title = "{The Phase II Murchison Widefield Array: Design overview}",
-      journal = {\pasa},
+      journal = {Publications of the Astronomical Society of Australia},
      keywords = {early universe, instrumentation: interferometers, radio continuum: general, radio lines: general, techniques: interferometric, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2018,
         month = nov,
@@ -257,7 +257,7 @@ archivePrefix = {arXiv},
          {Whitney}, A.~R. and {Williams}, A. and {Williams}, C.~L. and
          {Wyithe}, J.~S.~B.},
         title = "{The Murchison Widefield Array: The Square Kilometre Array Precursor at Low Radio Frequencies}",
-      journal = {\pasa},
+      journal = {Publications of the Astronomical Society of Australia},
      keywords = {instrumentation: interferometers, techniques: image processing, techniques: interferometric, radio continuum: general, radio lines: general, early Universe, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2013,
         month = jan,
