### 2020-07-13 #2363 @aureliocarnero
Dear @sambit-giri , I cannot find any reference in the github (or in the documentation) to:

So far I cannot evaluate this. I recommend you put this guidelines in the github repo README and possibly in the readthedocs.
To give you an example, you can see this other software I reviewed in the past, the section "Contributing":
https://github.com/noraeisner/LATTE#contributing
