+++ CONTRIBUTING.rst
@@ -1,0 +1,26 @@
+============
+Contributing
+============
+
+Thank you so much for using `tools21cm`. Contributions are welcome and are greatly appreciated! 
+
+Bug reports/Feedback/Questions
+===============================================
+It is very helpful to us when users report bugs or unexpected behaviour in the code. Please feel free to open an issue at the repository's [issue](https://github.com/sambit-giri/tools21cm/issues) page. It is also the best place to seek any help or would like to give us any feedback. 
+
+If you report an issue, please give the necessary details (python environment, code version, the unexpected output or error message) and a "Working Example" to reproduce the problem.
+
+Adding new features
+===================
+If you feel that there are some features missing, we welcome you to open an issue and discuss it. We can plan how this missing feature can be included in the code. 
+
+Setting up your development environment
+---------------------------------------
+In case you want to make substantial change to the code, you can fork the repository and clone it to your local machine.
+
+```
+git clone https://github.com/<YOURUSERNAME>/tools21cm.git
+cd tools21cm
+git checkout -b <BRANCHNAME>
+```
+where `<YOURUSERNAME>` and `<BRANCHNAME>` are your Github username and the name of new branch for local development respectively.
