+++ CONTRIBUTING.rst
@@ -6,7 +6,7 @@ Thank you so much for using `tools21cm`. Contributions are welcome and are great
 
 Bug reports/Feedback/Questions
 ===============================================
-It is very helpful to us when users report bugs or unexpected behaviour in the code. Please feel free to open an issue at the repository's [issue](https://github.com/sambit-giri/tools21cm/issues) page. It is also the best place to seek any help or would like to give us any feedback. 
+It is very helpful to us when users report bugs or unexpected behaviour in the code. Please feel free to open an issue at the repository's `issue <https://github.com/sambit-giri/tools21cm/issues>`_ page. It is also the best place to seek any help or would like to give us any feedback. 
 
 If you report an issue, please give the necessary details (python environment, code version, the unexpected output or error message) and a "Working Example" to reproduce the problem.
 
@@ -16,12 +16,11 @@ If you feel that there are some features missing, we welcome you to open an issu
 
 Setting up your development environment
 ---------------------------------------
-In case you want to make substantial change to the code, you can fork the repository and clone it to your local machine:
+In case you want to make substantial change to the code, you can fork the repository and clone it to your local machine::
 
-```bash
-git clone https://github.com/<YOURUSERNAME>/tools21cm.git
-cd tools21cm
-git checkout -b <BRANCHNAME>
-```
+       git clone https://github.com/<YOURUSERNAME>/tools21cm.git
+       cd tools21cm
+       git checkout -b <BRANCHNAME
 
-where `<YOURUSERNAME>` and `<BRANCHNAME>` are your Github username and the name of new branch for local development respectively.
+
+where ``<YOURUSERNAME>`` and ``<BRANCHNAME>`` are your Github username and the name of new branch for local development respectively.
