### 2020-07-14 #2363 @sambit-giri
> Dear @sambit-giri , I cannot find any reference in the github (or in the documentation) to:
>
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> So far I cannot evaluate this. I recommend you put this guidelines in the github repo README and possibly in the readthedocs.
> To give you an example, you can see this other software I reviewed in the past, the section "Contributing":
> https://github.com/noraeisner/LATTE#contributing

@aureliocarnero Thank you for the recommendation.
We have added a brief text about "contributing" in README.md and added the directed users to the detailed description in the documentation: https://tools21cm.readthedocs.io/contributing.html
