### 2020-07-15 #2363 @ziotom78
> In order to test the code, some simulation output is required. We have now provided some data at https://ttt.astro.su.se/~gmell/244Mpc/TestData_tools21cm/
> The link contains data along with a README explaining it.

Hi @sambit-giri , I have downloaded the data and checked that the plots produced by the tutorial match mines. They effectively look the same, but I interpret JOSS' requirement for "automated tests" in a stricter sense: it's the computer that should take care of checking the consistency of the results with their expected outcome. This ensures that patches from potential contributors do not introduce systematic effects or bugs in your codebase.

I realize that it's not feasible to include all the test data in the GitHub repository, as they are really too big; for me your solution of providing them in a separate URL is perfect. Perhaps you could produce smaller data files by sampling a tinier space? It's not required that the results of a tests are cosmologically interesting, they just need to check that modifications to the code do not introduce systematic effects. You could keep the tutorial as it is, relying on the bigger data files (the tutorial must be cosmologically interesting, after all!), and include the smaller files in the GitHub repository only for the automated tests.

Also, you could add unit tests to several functions, like `t2c.statistics.skewness`, without the need to load data from files. Example:

```python
# Put this in file test/test_statistics.py
from tools21cm.statistics import skewness
import numpy as np

# Test the functionality of "skewness"
assert np.allclose([
    skewness([1.0, 2.0, 0.0]),
    skewness([1.0, 2.0, 4.0],
)], [
    0.0,
    0.20782656212951636,
])
```

This kind of tests is very useful for documentation as well.
