### 2020-07-17 #2363 @aureliocarnero
Dear @sambit-giri I had an unexpected error running the tutorials in https://tools21cm.readthedocs.io/examples/tutorials.html#Bubble-size-distribution

When I execute cell [22]: r_spa, dn_spa = t2c.spa(xHII, boxsize=boxsize, nscales=20)
It tells me the following error:

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/acarnero/codes/anaconda3/lib/python3.7/site-packages/tools21cm/bubble_stats.py", line 181, in spa
    rs, ni = spa_np.spa_np(data, xth=xth, binning=binning, nscales=nscales)
  File "/home/acarnero/codes/anaconda3/lib/python3.7/site-packages/tools21cm/spa_np.py", line 15, in spa_np
    for i in xrange(nscales):
NameError: name 'xrange' is not defined

Basically, xrange is not defined...
Im running in python 3.
Cheers
Aurelio
