### 2020-07-17 #2363 @pibion
@sambit-giri I agree with @aureliocarnero, providing a separate URL for a data download is an excellent solution for larger data sets.

I would recommend putting this data on a service like Zenodo or Figshare rather than hosting it from your own website.
