+++ tests/test_PointStatistics.py
@@ -1,0 +1,7 @@
+import tools21cm as t2c
+import unittest
+
+data_flat  = np.ones((10,10,10))
+data_gauss = np.random.normal(0,1,size=(10,10,10))
+
+
