### 2020-07-21 #2363 @sambit-giri
Dear @ziotom78 and @aureliocarnero

I have put scripts for unit testing at https://github.com/sambit-giri/tools21cm/tree/master/tests
I have not included the functions that need data such as the read/write functions and redshift space distortion implementation.

Please let me know if anything else is needed.

Best,
Sambit
