### 2020-07-21 #2363 @aureliocarnero
Dear @sambit-giri , this is great, but how can I execute the unity tests? You should explain this in the documentation... Cheers
