### 2020-07-21 #2363 @sambit-giri
> Dear @sambit-giri , this is great, but how can I execute the unity tests? You should explain this in the documentation... Cheers

@aureliocarnero I have put a README inside the same folder.
