### 2020-07-21 #2363 @sambit-giri
> @sambit-giri I agree with @aureliocarnero, providing a separate URL for a data download is an excellent solution for larger data sets.
>
> I would recommend putting this data on a service like Zenodo or Figshare rather than hosting it from your own website.

Dear @pibion

We have put the data up on Zenodo (https://doi.org/10.5281/zenodo.3953639).
The link to the data has also been updated in the tutorial (https://tools21cm.readthedocs.io/examples/tutorials.html).

Best,
Sambit
