### 2020-07-22 bc576e2
subtract_mean_channelwise removed
