+++ tests/test_PowerSpectrum.py
@@ -31,6 +31,14 @@ def test_power_spectrum_1d():
 	slope  = (np.log10(pp*kk**3/2/np.pi**2)[kbins-3]-np.log10(pp*kk**3/2/np.pi**2)[3])/(np.log10(kk)[kbins-3]-np.log10(kk)[3])
 	assert np.abs(slope-3)<=0.1 
 
+def test_dimensionless_ps():
+	'''
+	With this test, power_spectrum_nd and radial_average are also test.
+	'''
+	dd, kk = t2c.dimensionless_ps(gauss, kbins=kbins, box_dims=box_dims)
+	slope  = (np.log10(dd)[kbins-3]-np.log10(dd)[3])/(np.log10(kk)[kbins-3]-np.log10(kk)[3])
+	assert np.abs(slope-3)<=0.25
+
 def test_power_spectrum_mu():
 	'''
 	With this test, power_spectrum_nd and mu_binning are also test.
