### 2020-07-22 b42c342
fixed docstrings
