### 2020-07-22 7c2bc61
fixed docstrings
