### 2020-07-22 499ea61
fixed docstrings
