### 2020-07-22 #2363 @sambit-giri
Dear @nicholebarry

> 1. subtract_mean_channelwise is in t2c.read_files. Is there a more logic place for this function?
I discovered that [subtract_mean_signal](https://tools21cm.readthedocs.io/contents.html#t2c.statistics.subtract_mean_signal) and subtract_mean_channelwise are doing the same thing. Therefore I have removed subtract_mean_channelwise. I think subtract_mean_signal is at the place.

> 2. There's a couple of places where the auto-generated descriptions on the readthedocs are hard to parse because there should be a newline and/or period. Some examples are in [Other Modules](https://tools21cm.readthedocs.io/contents_other.html) under "Some useful attributes" and on the [Main modules](https://tools21cm.readthedocs.io/contents.html) page under t2c.power_spectrum.power_spectrum_mu.

Thanks for pointing it out. I have fixed them.

> 3. In [Temperature](https://tools21cm.readthedocs.io/contents.html#module-t2c.temperature), please add units

Done.

> 4. In [Tutorials](https://tools21cm.readthedocs.io/examples/tutorials.html) under 21cm Brightness Temperature, please add units to the tutorial plot

Done.

> 5. In [Tutorials](https://tools21cm.readthedocs.io/examples/tutorials.html), please set x and y axis labels for all plots where applicable

Done.
