### 2020-07-22 #2363 @sambit-giri
@nicholebarry

> @sambit-giri
>
> Here are some more detailed comments/questions as I go through the tutorial.
>
> 1. In [Tutorials](https://tools21cm.readthedocs.io/examples/tutorials.html) under 21cm Brightness Temperature, it seems as though the sample data set already has its per-freq mean subtracted, and thus the subtract_mean_channelwise has very little effect. This may be confusing to some.

I don't think so. The [tutorial](https://tools21cm.readthedocs.io/examples/tutorials.html#21-cm-brightness-temperature) now prints the mean of the first channel before and after applying subtract_mean_signal.

> 2. In [Tutorials](https://tools21cm.readthedocs.io/examples/tutorials.html) under 21cm power spectrum, should the box_dims be the C2Ray box size (244) or the array size (250)? Looking at the code, it seems as though it should be 244, but this is confusing from a new-user tutorial standpoint. Add a comment?

I have commented on this in the [tutorial](https://tools21cm.readthedocs.io/examples/tutorials.html#21-cm-power-spectrum).

> 3. I tried to expand upon the tutorial a little bit in [21cm power spectrum](https://tools21cm.readthedocs.io/examples/tutorials.html#21-cm-power-spectrum), and I got the following error.
>
> ```
> t2c.power_spectrum.dimensionless_ps(dT_subtracted, kbins=15, box_dims=box_dims)
> Traceback (most recent call last):
>   File "<stdin>", line 1, in <module>
>   File "/Users/nabarry/opt/anaconda3/lib/python3.7/site-packages/tools21cm/power_spectrum.py", line 464, in dimensionless_ps
>     ks_new = np.array([ks[factor*(i+0.5)] for i in range(kbins)])
>   File "/Users/nabarry/opt/anaconda3/lib/python3.7/site-packages/tools21cm/power_spectrum.py", line 464, in <listcomp>
>     ks_new = np.array([ks[factor*(i+0.5)] for i in range(kbins)])
> IndexError: only integers, slices (`:`), ellipsis (`...`), numpy.newaxis (`None`) and integer or boolean arrays are valid indices
> ```
Thanks! I have fixed this error.
