### 2020-07-22 #2363 @aureliocarnero
Dear @sambit-giri could you please put in the documentation explicitly something about running the Automated tests? I know there is a redmine in the directory, but how a user can know that directory exist? Cheers
Aurelio
