### 2020-07-22 #2363 @nicholebarry
@sambit-giri

Thanks for implementing those comments.

Final comments on the tutorial:
1. Units for R and V in plots under bubble size distributions. Also titles to distinguish the plots using the various methods.
2. I don't want to delay the publication, so this is more of a suggestion for the future. There's a lot of fantastic modules for observational aspects (foregrounds, noise, beam kernels, etc) that you've implemented. I'd love to see a tutorial to highlight this functionality. I also think this would advertise to a wider potential user-base.

Misc. comments from browsing the modules:
1. t2c.telescope_functions.kelvin_2_jansky says it return muJy, but I think it may return Jy. Is this correct?

I look forward to using tools21cm in conjunction with 21cmfast in the future!
