### 2020-07-22 #2363 @sambit-giri
> @sambit-giri
>
> Thanks for implementing those comments.
>
> Final comments on the tutorial:
>
> 1. Units for R and V in plots under bubble size distributions. Also titles to distinguish the plots using the various methods.
> 2. I don't want to delay the publication, so this is more of a suggestion for the future. There's a lot of fantastic modules for observational aspects (foregrounds, noise, beam kernels, etc) that you've implemented. I'd love to see a tutorial to highlight this functionality. I also think this would advertise to a wider potential user-base.
>
> Misc. comments from browsing the modules:
>
> 1. t2c.telescope_functions.kelvin_2_jansky says it return muJy, but I think it may return Jy. Is this correct?
>
> I look forward to using tools21cm in conjunction with 21cmfast in the future!

Thanks @nicholebarry for the recommendations. The current plan is to publish the base version of tools21cm. Few modules for observational aspects, like the foreground models and the beam kernels, are still under development. Therefore we have not mentioned about it in the paper. As the thermal noise goes well under the hood of observational aspects, we decided to release them together in the next release version.
