### 2020-07-22 #2363 @sambit-giri
@nicholebarry

> @sambit-giri
>
> Thanks for implementing those comments.
>
> Final comments on the tutorial:
>
> 1. Units for R and V in plots under bubble size distributions. Also titles to distinguish the plots using the various methods.

Done.

> Misc. comments from browsing the modules:
>
> 1. t2c.telescope_functions.kelvin_2_jansky says it return muJy, but I think it may return Jy. Is this correct?
>

It is Jy. I have fixed it in the docstring.
