### 2020-07-23 302a58f
FOF always uses skimage
