### 2020-07-23 3c32978
thresholding with kmeans fixed
