### 2020-07-23 ee208fb
FOF always uses skimage
