### 2020-07-23 #2363 @sambit-giri
@nicholebarry

> I also ran into a few failures with unit testing (which look to be related to the dimensionless_ps bug seen above)
>
> ```
> ============================================================ FAILURES =============================================================
> ____________________________________________________ test_bubbles_from_kmeans _____________________________________________________
>
>     def test_bubbles_from_kmeans():
> >   	out = t2c.bubbles_from_kmeans(data_ball, n_clusters=2, upper_lim=False)
>
> test_IdentifyRegions.py:19:
> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
> ../../../opt/anaconda3/lib/python3.7/site-packages/tools21cm/identify_regions.py:63: in bubbles_from_kmeans
>     if n_clusters==2: array, t_th = threshold_kmeans(data, upper_lim=upper_lim, n_jobs=n_jobs)
> ../../../opt/anaconda3/lib/python3.7/site-packages/tools21cm/identify_regions.py:104: in threshold_kmeans
>     else: t_th = X[y==1].max()
> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
>
> a = array([], shape=(0, 1), dtype=uint8), axis = None, out = None, keepdims = False, initial = <no value>, where = True
>
>     def _amax(a, axis=None, out=None, keepdims=False,
>               initial=_NoValue, where=True):
> >       return umr_maximum(a, axis, None, out, keepdims, initial, where)
> E       ValueError: zero-size array to reduction operation maximum which has no identity
>
> ../../../opt/anaconda3/lib/python3.7/site-packages/numpy/core/_methods.py:30: ValueError
> ______________________________________________________ test_dimensionless_ps ______________________________________________________
>
>     def test_dimensionless_ps():
>     	'''
>     	With this test, power_spectrum_nd and radial_average are also test.
>     	'''
> >   	dd, kk = t2c.dimensionless_ps(gauss, kbins=kbins, box_dims=box_dims)
>
> test_PowerSpectrum.py:38:
> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
> ../../../opt/anaconda3/lib/python3.7/site-packages/tools21cm/power_spectrum.py:464: in dimensionless_ps
>     ks_new = np.array([ks[factor*(i+0.5)] for i in range(kbins)])
> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
>
> .0 = <range_iterator object at 0x11bc5dba0>
>
> >   ks_new = np.array([ks[factor*(i+0.5)] for i in range(kbins)])
> E   IndexError: only integers, slices (`:`), ellipsis (`...`), numpy.newaxis (`None`) and integer or boolean arrays are valid indices
>
> ../../../opt/anaconda3/lib/python3.7/site-packages/tools21cm/power_spectrum.py:464: IndexError
> ```

Did you git pull and install the package again?
I had completely rewritten the dimensionless_ps function after your previous comment about it. It doesn't have the above lines anymore (https://tools21cm.readthedocs.io/_modules/t2c/power_spectrum.html#dimensionless_ps).
I have also fixed the other error you got with your unit test.
