+++ docs/readme.rst
@@ -1,5 +1,5 @@
 =========
-tools21cm
+Tools21cm
 =========
 
 A python package for analysing 21-cm signals from the Epoch of Reionization (EoR) and Cosmic Dawn (CD). The source files can be found `here <https://github.com/sambit-giri/tools21cm>`_.
@@ -7,3 +7,7 @@ A python package for analysing 21-cm signals from the Epoch of Reionization (EoR
 Note: There are some modules in the package that are still under active development. Therefore please contact the authors if you get erronous results.
 
 
+Package details
+===============
+
+The package provides tools to analyse cosmological simulations of reionization and cosmic dawn. It contains modules to create mock 21-cm observations for current anf upcoming radio telescopes, such as LOFAR, MWA and SKA.
