### 2020-07-23 #2363 @sambit-giri
Dear @nicholebarry

> @sambit-giri
>
> > Thanks @nicholebarry for the recommendations. The current plan is to publish the base version of tools21cm. Few modules for observational aspects, like the foreground models and the beam kernels, are still under development. Therefore we have not mentioned about it in the paper. As the thermal noise goes well under the hood of observational aspects, we decided to release them together in the next release version.
>
> Makes total sense.
>
> I would suggest some more description/motivation on the front page of readthedocs (and subsequently the Github main readme) that would help a first-time user know whether or not this package satisfies their needs. Feel free to use these suggestions (or better ones you come up with). Just a few sentences for each would do. For reference, I'm using [pyuvdata](https://pyuvdata.readthedocs.io/en/latest/) as a guide.
>
> 1. Inputs: the simulation codes that you can currently input into tools21cm
> 2. Outputs: bubble sizes, power spectrum, redshift distortions, etc
> 3. Under-development: instrumental effects, noise, other specific simulation code inputs
> 4. Citation: point to JOSS paper when published and a snippet about how people should cite the code.
> 5. Tests: description on how to run the unit tests (also great for people to know if they've installed correctly).
>

We have updated the documentation as per your recommendations. We will put the "Citation" after a DOI is produced.

Best,
Sambit
