+++ docs/installation.rst
@@ -35,10 +35,7 @@ Or look `here <http://docs.python.org/2/install/>`_ for more details.
 
 Tests
 -----
-For testing, one can use `pytest <https://docs.pytest.org/en/stable/>`_ or `nosetests <https://nose.readthedocs.io/en/latest/>`_. Both packages can be installed using pip. To run all the test script, run the either of the following:
-
-.. code-block:: python
-   :linenos:
+For testing, one can use `pytest <https://docs.pytest.org/en/stable/>`_ or `nosetests <https://nose.readthedocs.io/en/latest/>`_. Both packages can be installed using pip. To run all the test script, run the either of the following::
 
     pytest -v 
-	nosetests -v
+    nosetests -v
