### 2020-07-27 sambit-giri/tools21cm#7 @sambit-giri
@ziotom78 Thanks for pointing this out.

I was unable to reproduce this error. With a quick search, I found that this could be caused when pytest is installed outside of a virtual environment where the package is installed (https://medium.com/@dirk.avery/pytest-modulenotfounderror-no-module-named-requests-a770e6926ac5). I did it and got the error.

The fix you suggested is working in that case. I will add a comment about this in the README.
