### 2020-07-27 sambit-giri/tools21cm#7 @ziotom78
> With a quick search, I found that this could be caused when pytest is installed outside of a virtual environment where the package is installed (https://medium.com/@dirk.avery/pytest-modulenotfounderror-no-module-named-requests-a770e6926ac5). I did it and got the error.

Yes, this was indeed the case, but then I installed `pytest` under my virtualenv, downloaded and installed `tools21cm` again and the error kept popping out. I believe it's safer for the general user if you provide the `python -m ptyest` command.
