### 2020-07-27 sambit-giri/tools21cm#7 @sambit-giri
Okay, I will do that.

While testing from the package directory, the command should be `python -m pytest tests/` and inside the `tests` directory, it should `python -m pytest`. Is this correct?
