### 2020-07-27 sambit-giri/tools21cm#7 @ziotom78
> While testing from the package directory, the command should be python -m pytest tests/ and inside the tests directory, it should python -m pytest. Is this correct?

Right, I can confirm it works in both ways.
