+++ docs/installation.rst
@@ -37,5 +37,5 @@ Tests
 -----
 For testing, one can use `pytest <https://docs.pytest.org/en/stable/>`_ or `nosetests <https://nose.readthedocs.io/en/latest/>`_. Both packages can be installed using pip. To run all the test script, run the either of the following::
 
-    pytest -v 
+    python -m pytest tests 
     nosetests -v
