### 2020-07-27 sambit-giri/tools21cm#7 @sambit-giri
The README and the instructions in the readthedocs documentation have been updated.
