### 2020-07-27 #2363 @ziotom78
Ok, [issue #7](https://github.com/sambit-giri/tools21cm/issues/7) is closed, my review is complete. Thank you all!
