### 2020-07-31 #2363 @pibion
@sambit-giri we have just a few more steps - I need to do a read-through of the final pdf and then you'll need to archive your software on Zenodo or Figshare.  I'll be able to review the pdf tomorrow and will ping you on the action you need to take tomorrow.
