### 2020-08-05 #2363 @pibion
@sambit-giri the pdf looks great.  I have one requested change: the last sentence in the second paragraph isn't immediately clear.

I would recommend splitting this sentence into two to make it clearer that Tools21cm takes the simulation outputs as its inputs.  Maybe something like, " Much of the functionality of Tools21cm is focused
on the construction and analysis of mock 21 cm observations in the context of current and
upcoming radio telescopes, such as the Low Frequency Array (LOFAR; Haarlem et al., 2013),
the Murchison Widefield Array (MWA; Tingay et al., 2013; Wayth et al., 2018), Hydrogen
Epoch of Reionization Array (HERA; DeBoer et al., 2017) and the Square Kilometre Array
(SKA; Mellema et al., 2013).  Tools21cm takes the simulation data of these telescopes as its inputs."
