+++ t2c/topology.py
@@ -23,7 +23,9 @@ def EulerCharacteristic(data, thres=0.5, neighbors=6):
 	Euler characteristics value.
 	"""
 	A = 1*(data>thres)
-	if 'numba' in sys.modules: from . import ViteBetti_numba as VB
+	if 'numba' in sys.modules: 
+		print('Using numba to speed up.')
+		from . import ViteBetti_numba as VB
 	else: from . import ViteBetti as VB
 	if neighbors==6 or neighbors==4: C = VB.CubeMap(A)
 	else: C = VB.CubeMap(1-A)
