+++ t2c/topology.py
@@ -25,13 +25,13 @@ def EulerCharacteristic(data, thres=0.5, neighbors=6, use_numba=False):
 	Euler characteristics value.
 	"""
 	A = 1*(data>thres)
-	# if 'numba' in sys.modules: 
-	# 	print('Using numba to speed up.')
-	# 	from . import ViteBetti_numba as VB
-	# else: from . import ViteBetti as VB
+	#if 'numba' in sys.modules: 
+	#	print('Using numba to speed up.')
+	#	from . import ViteBetti_numba as VB
+	#else: from . import ViteBetti as VB
 	if use_numba: 
 		print('Using numba to speed up.')
-	 	from . import ViteBetti_numba as VB
+		from . import ViteBetti_numba as VB
 	else: 
 		from . import ViteBetti as VB
 	if neighbors==6 or neighbors==4: C = VB.CubeMap(A)
