### 2020-08-05 #2363 @pibion
@sambit-giri this looks excellent, thank you!

I'll need you to make a tagged release and archive, and report the version number and archive DOI in this review thread.

Zenodo, Figshare, and OSF are good options.  I like to use Zenodo because you can set it up to automatically create a new DOI every time you add a new tag to your repository.
