### 2020-08-07 #2363 @danielskatz
@sambit-giri - when "21 cm" is used as an adjective, as in "21 cm signal", it should be hyphenated to "21-cm signal".  Can you change this in your paper and in the title?  You might also want to change the title of the zenodo archive to match.  Actually, I now see that you need to change the title of the Zenodo archive in any event, as it's the default title from the repo.
