### 2020-08-10 #2363 @sambit-giri
Dear @danielskatz

I have edited "21 cm" in the paper wherever needed. I have also changed the title of the zenodo archive.

Best,
Sambit
